refactor: remove contact field and associated validation schemas from organizations.ts to organization.validation.ts
